fix: QueryDsl 결과와 record 의 파라미터 타입이 맞지 않는 버그 수정

+++ src/main/java/org/example/mopl/content/repository/ContentQueryRepository.java
@@ -364,8 +364,8 @@ public class ContentQueryRepository {
   }
 
   private record ReviewStat(
-      Long sum,
-      Integer count
+      Double sum,
+      Long count
   ) {
 
   }
